+++ Runtime/Contexts/ADependencyInjectionContext.cs
@@ -7,7 +7,7 @@ using UnityEngine.SceneManagement;
 namespace JakubKrizanovsky.DependencyInjection
 {
     [DefaultExecutionOrder(-1000)]
-    public class ADependencyInjectorContext : MonoBehaviour, IDependencyInjectorContext
+    public class ADependencyInjectorContext : MonoBehaviour
     {
         private readonly Dictionary<Type, object> _services = new();
 
@@ -30,12 +30,23 @@ namespace JakubKrizanovsky.DependencyInjection
                 foreach(MonoBehaviour component in components) {
                     Type type = component.GetType();
                     if(Attribute.IsDefined(type, typeof(ServiceAttribute))) {
-                        RegisterService(type, component);
-                        HandleServicePersistence(component, type);
+                        ServiceAttribute attribute = type.GetCustomAttribute<ServiceAttribute>();
+                        if(!attribute.Persistent) {
+                            RegisterService(type, component);
+                        } else {
+                            // Persistent services are always registered as global
+                            DependencyInjector.RegisterService(component, true);
+                        }
                     }
 
                     if(component is IServiceProvider serviceProvider) {
-                        RegisterService(serviceProvider.Service);
+                        object service = serviceProvider.Service;
+                        if(serviceProvider.Persistent) {
+                            DependencyInjector.RegisterService(service, true);
+                            MakePersistent(service);
+                        } else {
+                            RegisterService(service);
+                        }
                     }
 
                     if(InjectableMembersCache.GetInjectableMembers(type).HasAny) {
@@ -50,14 +61,14 @@ namespace JakubKrizanovsky.DependencyInjection
             }
         }
 
-        public void RegisterService<T>(T service) {
+        internal void RegisterService<T>(T service) {
             RegisterService(service.GetType(), service);
         }
 
-        public void RegisterService(Type type, object service) {
+        internal virtual void RegisterService(Type type, object service) {
             ServiceAttribute attribute = type.GetCustomAttribute<ServiceAttribute>();
 
-            // Make sure unique services only get 
+            // Make sure unique services only get registered once
             if(attribute != null && attribute.Unique && _services.ContainsKey(type)) {
                 if(service is MonoBehaviour serviceMB) {
                     Destroy(serviceMB);
@@ -73,15 +84,14 @@ namespace JakubKrizanovsky.DependencyInjection
             }
         }
 
-        public bool TryResolve(Type type, out object service) {
+        internal bool TryResolve(Type type, out object service) {
             return _services.TryGetValue(type, out service);
         }
 
-        private void HandleServicePersistence(MonoBehaviour service, Type type) {
-            ServiceAttribute attribute = type.GetCustomAttribute<ServiceAttribute>();
-            if(attribute.Persistent) {
-                service.transform.parent = null;
-                DontDestroyOnLoad(service.gameObject);
+        protected void MakePersistent(object service) {
+            if(service is MonoBehaviour serviceMB) {
+                serviceMB.transform.parent = null;
+                DontDestroyOnLoad(serviceMB.gameObject);
             }
         }
 	}