+++ Runtime/Contexts/SceneDependencyInjectionContext.cs
@@ -3,7 +3,6 @@ using UnityEngine;
 namespace JakubKrizanovsky.DependencyInjection
 {
     [DefaultExecutionOrder(-1000)]
-    [Service]
     public class SceneDependencyInjectionContext : ADependencyInjectorContext
     {