+++ Runtime/Support/IServiceProvider.cs
@@ -3,5 +3,6 @@ namespace JakubKrizanovsky.DependencyInjection
     public interface IServiceProvider
     {
         public object Service {get;}
+        public bool Persistent => false;
     }
 }