+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 ## [1.1.1] - 2026-04-23
 ### Fix
 - Fix discovery when returning to global context scene
+
+## [1.2.0] - 2026-04-24
+### Improvements and Refactor
+- Persistent services are now always registered at the global context
+- ServiceProxy can now make services persistent
+- Some minor API changes, manual service registration now done directly with DependencyInjector